pid_calibrate: Improved calibration via an adaptive algorithm
I updated the pid calibration algorithm based on the processes outlined
in this paper. https://www.osti.gov/biblio/1646635

The update minimizes two undesirable phenomena, bias and asymmetry.
Tests of my own and others can be seen in this klipper discord thread
https://klipper.discourse.group/t/experimental-pid-improvement-changes/3604

I also updated the code to log Cohen-Coon calibration constants, because
They can be calculated with the data on hand and work better for high
thermal mass beds found in a lot of higher end printers.

Signed-off-by: Daniel Sherman <blogmail79@dans-hobbies.com>